+++ src/games/tf2.rs
@@ -71,10 +71,10 @@ impl Response {
 }
 
 pub fn query(address: &str, port: Option<u16>) -> GDResult<Response> {
-    let valve_response = ValveProtocol::query(App::TF2, address, match port {
+    let valve_response = ValveProtocol::query(address, match port {
         None => 27015,
         Some(port) => port
-    }, None)?;
+    }, Some(App::TF2), None)?;
 
     Ok(Response::new_from_valve_response(valve_response))
 }