+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Who knows what the future holds...
 
 # 0.0.4 - ??/??/????
 Queries now support DNS resolve.  
+Changed Valve Protocol 3rd argument to Option<GatherSettings>, being None means everything should be gathered.  
+Better appid unknown cast error.  
 [Alien Swarm](https://store.steampowered.com/app/630/Alien_Swarm/) implementation (not tested).  
 [Alien Swarm: Reactive Drop](https://store.steampowered.com/app/563560/Alien_Swarm_Reactive_Drop/) implementation.  
 [Insurgency](https://store.steampowered.com/app/222880/Insurgency/) implementation.